+++ src/ripple_detection/detectors/__init__.py
@@ -12,6 +12,7 @@ from ripple_detection.detectors._lfp import (
     get_Yu_ripple_consensus_trace,
 )
 from ripple_detection.detectors._long import Long_sharp_wave_ripple_detector
+from ripple_detection.detectors._trace import detect_events_from_trace
 from ripple_detection.detectors._zugaro import Zugaro_ripple_detector
 
 __all__ = [
@@ -23,6 +24,7 @@ __all__ = [
     "Shvartsman_ripple_detector",
     "Yu_ripple_detector",
     "Zugaro_ripple_detector",
+    "detect_events_from_trace",
     "get_Kay_ripple_consensus_trace",
     "get_Yu_ripple_consensus_trace",
     "multiunit_HSE_detector",